Generate HTML from OmniOutliner document.
The real reason for this script is that OmniOutliner refuses to export
to HTML to a directory name that does not end in ".html".  I find this
annoying and confusing.  Thankfully, it turns out that when the export
command is invoked via AppleScript, it is possible to specify a
destination folder/directory without ending it with the suffix
".html".

This script can be invoked from a shell and passing it a destination
path like this:

   osascript export-oo-as-html.scpt /some/path/to/a/directory